move context -> /pkg/auth/context
