+++ pkg/auth/auth_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
 
-	"github.com/transcom/mymove/pkg/context"
+	"github.com/transcom/mymove/pkg/auth/context"
 )
 
 type AuthSuite struct {
